refactored classes names and readme file
